scripts/requirements.txt
requests>=2.31
python-dotenv>=1.0
pillow>=10.0
numpy>=1.26
openpyxl>=3.1
mutagen>=1.47
